refactor: split class.c into src/meta/class/ modules (#114)
src/meta/class.c held the class build end to end: the create path with the
hand-off gate, the settle machinery, and the field installation. This
splits it along those seams, moving the code verbatim, following the #68
precedent:

  src/meta/class/create.c   build_struct_class, the hand-off gate, the
                            keyword probe, create_class
  src/meta/class/settle.c   the rebind tables, the settle verify and
                            rebinds, settle_cache_fill, the C pins
  src/meta/class/install.c  install_fields, slot/member offset
                            resolution, the constructor wiring

Each function keeps its old linkage: file-local helpers stay static in
their file, and only the six functions that cross a boundary
(install_fields, install_post_init, defines_own_init,
verify_settle_names_readable, settle_planned, settle_mro_bindings) are
extern, declared in meta.h, which now forward-declares struct field_plan
rather than including fields.h (the pointer-only declarations do not need
the definition, and the include would arrive before StructType). The
TESTING section lives with the settle machinery it instruments, so its two
observability statics stay file-local.

No behaviour change: the matrix gate (3.10-3.15, free-threaded, the
in-file C tests, the analyzers) passes on the split as on the original.
