+++ build_config.py
@@ -19,7 +19,9 @@ BUILD: Final = BuildConfig(
         "src/meta/meta.c",
         "src/meta/bases.c",
         "src/meta/namespace.c",
-        "src/meta/class.c",
+        "src/meta/class/create.c",
+        "src/meta/class/install.c",
+        "src/meta/class/settle.c",
         "src/mixin.c",
         "src/options.c",
         "src/repr.c",
