+++ src/meta/class/install.c
@@ -1,0 +1,237 @@
+#include <Python.h>
+#include <stdbool.h>
+
+#include "../../construct.h"
+#include "../../fields.h"
+#include "../meta.h"
+#include "../../owned.h"
+#include "../../types.h"
+
+static Py_ssize_t * resolve_slot_offsets(
+	StructType * struct_class,
+	StructType const * base,
+	PyObject * new_names,
+	Py_ssize_t field_count
+);
+static Py_ssize_t * resolve_member_offsets(
+	StructType * struct_class,
+	Py_ssize_t const * slot_offsets,
+	Py_ssize_t field_count,
+	Py_ssize_t * member_count
+);
+
+enum result install_fields(
+	StructType * const struct_class,
+	StructType const * const base,
+	struct field_plan const * const plan,
+	struct options const options,
+	bool const resolves_body_eq
+) {
+	PY_MOVABLE(field_names, PyList_AsTuple(plan->all_names));
+
+	if (field_names == NULL) {
+		return RESULT_ERROR;
+	}
+
+	Py_ssize_t const field_count = PyTuple_GET_SIZE(field_names);
+	Py_ssize_t * const offsets =
+		resolve_slot_offsets(struct_class, base, plan->new_names, field_count);
+
+	if (offsets == NULL) {
+		return RESULT_ERROR;
+	}
+
+	Py_ssize_t member_count = 0;
+	Py_ssize_t * const member_offsets =
+		resolve_member_offsets(struct_class, offsets, field_count, &member_count);
+
+	if (member_offsets == NULL) {
+		PyMem_Free(offsets);
+
+		return RESULT_ERROR;
+	}
+
+	struct_class->struct_field_names = py_move(&field_names);
+	struct_class->struct_defaults = Py_NewRef(plan->defaults);
+	struct_class->struct_slot_offsets = offsets;
+	struct_class->struct_member_offsets = member_offsets;
+	struct_class->struct_member_count = member_count;
+	struct_class->struct_field_count = field_count;
+	struct_class->struct_default_count = PyTuple_GET_SIZE(plan->defaults);
+	struct_class->struct_options = options;
+	struct_class->struct_resolves_body_eq = resolves_body_eq;
+
+	if (defines_own_init(struct_class)) {
+		struct_class->heap_type.ht_type.tp_new = Struct_new;
+	} else {
+		struct_class->heap_type.ht_type.tp_vectorcall = Struct_vectorcall;
+	}
+
+	return install_post_init(struct_class);
+}
+
+bool defines_own_init(StructType const * const struct_class) {
+	return struct_class->heap_type.ht_type.tp_init != PyBaseObject_Type.tp_init;
+}
+
+enum result install_post_init(StructType * const struct_class) {
+	PyObject * const hook = optional_attribute((PyObject *) struct_class, "__post_init__");
+
+	if (hook == NULL && PyErr_Occurred()) {
+		return RESULT_ERROR;
+	}
+
+	Py_XSETREF(struct_class->struct_post_init, hook);
+
+	return RESULT_OK;
+}
+
+static Py_ssize_t * resolve_slot_offsets(
+	StructType * const struct_class,
+	StructType const * const base,
+	PyObject * const new_names,
+	Py_ssize_t const field_count
+) {
+	Py_ssize_t * const offsets = PyMem_New(Py_ssize_t, field_count > 0 ? field_count : 1);
+
+	if (offsets == NULL) {
+		PyErr_NoMemory();
+
+		return NULL;
+	}
+
+	Py_ssize_t const inherited_count = base != NULL ? base->struct_field_count : 0;
+
+	for (Py_ssize_t i = 0; i < inherited_count; ++i) {
+		offsets[i] = base->struct_slot_offsets[i];
+	}
+
+	PyMemberDef const * const members = struct_heap_type_members(struct_class);
+	Py_ssize_t const member_count = Py_SIZE(struct_class);
+
+	for (Py_ssize_t i = 0; i < PyList_GET_SIZE(new_names); ++i) {
+		PyObject * const field_name = PyList_GET_ITEM(new_names, i);
+		struct member_lookup const found = find_member(members, member_count, field_name);
+
+		switch (found.tag) {
+			case MEMBER_LOOKUP_ERROR:
+				PyMem_Free(offsets);
+				return NULL;
+			case MEMBER_LOOKUP_MISSING:
+				PyErr_Format(PyExc_RuntimeError, "could not find slot offset for %R", field_name);
+				PyMem_Free(offsets);
+
+				return NULL;
+			case MEMBER_LOOKUP_FOUND:
+				offsets[inherited_count + i] = found.slot_offset;
+		}
+	}
+
+	return offsets;
+}
+
+static Py_ssize_t * resolve_member_offsets(
+	StructType * const struct_class,
+	Py_ssize_t const * const slot_offsets,
+	Py_ssize_t const field_count,
+	Py_ssize_t * const member_count
+) {
+	/* A non-struct base's own __slots__ members are not struct fields, so
+	 * the copy would never touch them without this table: one walk here, at
+	 * class creation, instead of rescanning every MRO dict on every copy.
+	 * The struct's own field descriptors are skipped by offset, and a
+	 * weakref slot is a getset descriptor, never a member one. */
+	*member_count = 0;
+
+	PyObject * const mro = struct_class->heap_type.ht_type.tp_mro;
+
+	for (Py_ssize_t i = 0; i < PyTuple_GET_SIZE(mro); ++i) {
+		PyTypeObject * const entry = (PyTypeObject *) PyTuple_GET_ITEM(mro, i);
+		PY_OWNED(entry_dict, struct_type_dict(entry));
+
+		if (entry_dict == NULL) {
+			return NULL;
+		}
+
+		Py_ssize_t position = 0;
+		PyObject * key;
+		PyObject * value;
+
+		while (PyDict_Next(entry_dict, &position, &key, &value)) {
+			if (!PyObject_TypeCheck(value, &PyMemberDescr_Type)) {
+				continue;
+			}
+
+			PyMemberDef const * const member = ((PyMemberDescrObject *) value)->d_member;
+
+			if (member == NULL || member->type != SLOT_MEMBER_TYPE) {
+				continue;
+			}
+
+			bool is_struct_field = false;
+
+			for (Py_ssize_t f = 0; f < field_count; ++f) {
+				if (member->offset == slot_offsets[f]) {
+					is_struct_field = true;
+					break;
+				}
+			}
+
+			if (!is_struct_field) {
+				++*member_count;
+			}
+		}
+	}
+
+	Py_ssize_t * const offsets = PyMem_New(Py_ssize_t, *member_count > 0 ? *member_count : 1);
+
+	if (offsets == NULL) {
+		PyErr_NoMemory();
+
+		return NULL;
+	}
+
+	Py_ssize_t written = 0;
+
+	for (Py_ssize_t i = 0; i < PyTuple_GET_SIZE(mro); ++i) {
+		PyTypeObject * const entry = (PyTypeObject *) PyTuple_GET_ITEM(mro, i);
+		PY_OWNED(entry_dict, struct_type_dict(entry));
+
+		if (entry_dict == NULL) {
+			PyMem_Free(offsets);
+
+			return NULL;
+		}
+
+		Py_ssize_t position = 0;
+		PyObject * key;
+		PyObject * value;
+
+		while (PyDict_Next(entry_dict, &position, &key, &value)) {
+			if (!PyObject_TypeCheck(value, &PyMemberDescr_Type)) {
+				continue;
+			}
+
+			PyMemberDef const * const member = ((PyMemberDescrObject *) value)->d_member;
+
+			if (member == NULL || member->type != SLOT_MEMBER_TYPE) {
+				continue;
+			}
+
+			bool is_struct_field = false;
+
+			for (Py_ssize_t f = 0; f < field_count; ++f) {
+				if (member->offset == slot_offsets[f]) {
+					is_struct_field = true;
+					break;
+				}
+			}
+
+			if (!is_struct_field) {
+				offsets[written++] = member->offset;
+			}
+		}
+	}
+
+	return offsets;
+}
