+++ src/meta/meta.h
@@ -7,6 +7,7 @@
 #include "../result.h"
 
 typedef struct StructType StructType;
+struct field_plan;
 
 enum hash_binding {
 	HASH_BODY_DEFINED,
@@ -144,3 +145,34 @@ PyObject * build_struct_class(
 	PyObject * keywords,
 	struct salix_state * state
 );
+enum result settle_planned(
+	StructType * struct_class,
+	StructType const * base,
+	PyObject * bases,
+	PyObject * name,
+	struct field_plan const * plan,
+	PyObject * original_namespace,
+	struct options options,
+	struct options inherited,
+	bool frozen_across_bases,
+	bool body_defines_eq,
+	bool inherits_body_eq,
+	bool derive_not_equal
+);
+enum result settle_mro_bindings(
+	StructType * struct_class,
+	PyObject * bases,
+	PyObject * original_namespace,
+	struct binding_plan bindings,
+	struct options options
+);
+enum result verify_settle_names_readable(PyObject * original_namespace);
+enum result install_fields(
+	StructType * struct_class,
+	StructType const * base,
+	struct field_plan const * plan,
+	struct options options,
+	bool resolves_body_eq
+);
+enum result install_post_init(StructType * struct_class);
+bool defines_own_init(StructType const * struct_class);
